requirements.txt
argcomplete==1.10.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
compressed-rtf==1.0.6
cryptography==44.0.1
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
idna==3.10
IMAPClient==2.1.0
lxml==5.3.1
numpy==1.24.4
olefile==0.47
pandas==2.0.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
pycparser==2.22
pycryptodome==3.21.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==0.6.23
pytz==2025.1
rapidfuzz==3.9.7
requests==2.32.3
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
textract==1.6.5
thefuzz==0.22.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.2.3
xlrd==1.2.0
XlsxWriter==3.2.2
